# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HMCA API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pms/amenities": {
            "get": {
                "tags": [
                    "Amenity"
                ],
                "summary": "Get all amenities",
                "operationId": "e1b4dd3820f1c4551464b623c2905578",
                "responses": {
                    "200": {
                        "description": "List of amenities"
                    }
                }
            },
            "post": {
                "tags": [
                    "Amenity"
                ],
                "summary": "Create a new amenity",
                "operationId": "60111e745d5803aa3f93eb1101e407da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_code",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Wi-Fi"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Amenity created successfully"
                    }
                }
            }
        },
        "/api/pms/amenities/{id}": {
            "get": {
                "tags": [
                    "Amenity"
                ],
                "summary": "Get amenity by ID",
                "operationId": "f15b4a58dbf3bd9184aa863c026e6382",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amenity details fetched successfully"
                    }
                }
            },
            "put": {
                "tags": [
                    "Amenity"
                ],
                "summary": "Update an existing amenity",
                "operationId": "da0686192e68b94659c32e63f89617e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Wi-Fi"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Amenity updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Amenity"
                ],
                "summary": "Delete an amenity",
                "operationId": "187681c2c553a290ab41744a221d90ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amenity deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}